+++ backend/apps/integrations/models/__init__.py
@@ -4,6 +4,7 @@ from apps.integrations.models.whatsapp_settings import WhatsAppSettings, WhatsAp
 from apps.integrations.models.telephony_agent import TelephonyAgent
 from apps.integrations.models.lead_sync_source import LeadSyncSource
 from apps.integrations.models.email_account import EmailAccount
+from apps.integrations.models.mail_domain import MailDomain
 from apps.integrations.models.email_message import EmailAttachment, EmailMessage
 from apps.integrations.models.email_campaign import EmailTemplate, EmailCampaign, EmailCampaignLog
 
@@ -15,6 +16,7 @@ __all__ = [
     "TelephonyAgent",
     "LeadSyncSource",
     "EmailAccount",
+    "MailDomain",
     "EmailMessage",
     "EmailAttachment",
     "EmailTemplate",